requirements.txt
langchain
langchain-openai
langchain-pinecone
pyodbc
python-dotenv
pandas
